Remove dist.sh and scrub.sh
These files seem to be remnants from when CVS and old versions of
automake were used.

Instead of dist.sh, "make distcheck" should be used.

Instead of scrub.sh, "git clean -dfx" or a similar command can be used.

Thanks to Johnny A. Solbu <johnny@solbu.net> for bringing this up and
suggesting the use of "git clean -dfx" instead of "./scrub.sh".