Run as oidentd:oidentd by default.
Only if the oidentd user does not exist, fall back to running as nobody,
or UID 65534, in this order.

Only if the oidentd group does not exist, fall back to running as nobody,
nogroup, or GID 65534, in this order.

Thanks to Johannes Löthberg <johannes@kyriasis.com> for this suggestion.