Implicitly drop privileges only when running as root
The previous behavior was to always drop privileges, including when
the --user / --group options were not given and oidentd was already
running as a non-root user.

This commit changes the behavior so that setuid() is only called when
either the UID or the EUID is zero, and so that setgid() is only
called when either the GID or the EGID is zero.