freebsd5: replace ucred with xucred
FreeBSD 5 introduced the `xucred` struct for external users in commit
c0511d3b58c985e17c0c23585ea7d16b6f52fe61 to avoid breaking user space
applications when the `ucred` struct changes.

oidentd happened to still work after this commit because the EUID it
reads was stored at the same location in both structures. This is no
longer the case with FreeBSD 13.

This commit updates the freebsd5 module to use `xucred` instead of
`ucred` as the expected return type for getcred calls.